chore: refactor role object
